cdk_stack/pyproject.toml
[tool.poetry]
name = "stack"
version = "0.1.0"
description = ""
authors = ["Dino Hensen <dino.hensen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "^1.76.0"
"aws-cdk.aws-ec2" = "^1.76.0"
"aws-cdk.aws-ecs" = "^1.76.0"
"aws-cdk.aws-elasticloadbalancingv2" = "^1.76.0"
"aws-cdk.aws-ecs-patterns" = "^1.76.0"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
